+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Fail and log tracker creation error when SLA Date field is missing (OSIDB-4856)
+- Adjust Jira tracker sync to use `Update Stream` field and `ps_component` label (OSIDB-3830)
 
 ## [5.8.0] - 2026-03-19
 ### Changed